pyproject.toml
[tool.poetry]
name = "potree_backend"
version = "0.0.1"
description = "Ease Casino"
authors = ["Djoncarton1921 <shapranov.nik@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.12"
redis = "^4.3.4"
celery = "^5.2.7"
Django = "^4.1.1"
django-filter = "^22.1"
django-phonenumber-field = "^7.0.0"
django-inlinecss = "^0.3.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
PyYAML = "^6.0"
pandas = "^1.4.4"
channels-redis = "^3.4.1"
python-decouple = "^3.6"
Pillow = "^9.2.0"
phonenumberslite = "^8.12.55"
djangochannelsrestframework = "^1.1.0"
django-rest-fuzzysearch = "^0.5.1"
mypy = "^0.971"
django-storages = "^1.13.1"
uWSGI = "^2.0.20"
django-cors-headers = "^3.13.0"
django-split-settings = "^1.2.0"
django-minio-backend = "^3.3.2"
isort = "^5.10.1"
flake8 = "^5.0.4"
black = "^22.10.0"
python-gettext = "^4.1"
drf-standardized-errors = "^0.12.3"
django-cryptocurrency-payment = "^0.1.2"
django-money = "^3.2.0"
ccxt = "^4.0.105"
psycopg2-binary = "^2.9.8"
hdwallet = "^2.2.1"
blockcypher = {git = "https://github.com/Shapranov-Maxim/blockcypher-python.git"}
laspy = "1.7.0"
boto3 = "^1.34.149"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
